moved seeding and db session logic under db folder
