+++ src/artifactminer/db/__init__.py
@@ -1,0 +1,15 @@
+"""Database module exposing models, session management, and utilities."""
+
+from .database import Base, engine, SessionLocal, get_db
+from .models import Artifact, Question
+from .seed import seed_questions
+
+__all__ = [
+    "Base",
+    "engine", 
+    "SessionLocal",
+    "get_db",
+    "Artifact",
+    "Question",
+    "seed_questions",
+]
